models/multirocket-new/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multirocket-new"
version = "0.1.0"
description = "Kubeflow-native MR-HY-SP workflows for C-MAPSS experiments"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "aeon==0.9.0",
  "h5py==3.11.0",
  "kfp[kubernetes]==2.15.0",
  "kubernetes>=29,<31",
  "numpy==1.24.4",
  "pandas==2.0.3",
  "pydantic>=2.7,<3",
  "PyYAML==6.0.2",
  "scikit-learn==1.3.2",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.5",
]
train = [
  "torch==2.4.1",
]

[project.scripts]
multirocket-new = "multirocket_new.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
